requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
qdrant-client==1.11.3
datasets==2.21.0
pyarrow==17.0.0
numpy==1.26.4
pydantic==2.8.2
python-dotenv==1.0.1
sentence-transformers==5.2.3
requests==2.32.3
